requirements.txt
Django>=3.1.0,<3.2.0
djangorestframework>-3.11.0,<3.12.0
psycopg2>=2.8.5,<2.9.0